requirements.txt
# DSA Tutor RAG — Complete Setup

streamlit==1.50.0
python-dotenv==1.1.1
groq==0.33.0
ollama==0.6.0
pinecone==7.3.0
pinecone-plugin-interface==0.0.7
pinecone-plugin-assistant==1.8.0
sentence-transformers==5.1.1
scikit-learn==1.7.2
torch==2.8.0
transformers==4.57.0
faiss-cpu==1.12.0
requests==2.32.5
tqdm==4.67.1
jsonlines==4.0.0
pandas==2.3.3
numpy==2.2.6
plotly==6.3.1
matplotlib==3.10.7
typing_extensions==4.15.0
python-dateutil==2.9.0.post0
ipykernel
jupyterlab
notebook
pytest
black
flake8
